+++ shoutout/MKMapView+ZoomLevel.m
@@ -149,7 +149,7 @@
 	return region;
 }
 
-- (NSUInteger) zoomLevel {
+- (double) zoomLevel {
     MKCoordinateRegion region = self.region;
     
     double centerPixelX = [MKMapView longitudeToPixelSpaceX: region.center.longitude];
